Feature: update config API support raw yaml payload
